getnet_agent_linux/requirements.txt
# Dependencias del Agente Getnet Linux

# Framework web
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Comunicación serial con POS
pyserial==3.5

# Variables de entorno
python-dotenv==1.0.0

# Validación de datos
pydantic==2.5.0



